fix: resolve base64ct version issue and improve sc-network compatibility
- Fix base64ct version from 1.8.0 to 1.6.0 to resolve security audit failures
- Add comprehensive sc-network-types patch for kad module compatibility
- Fix missing RootOrigin in ismp_grandpa::Config implementation
- Comment out problematic pallet_network_monitor import in cere-dev runtime
- Update Cargo.lock with compatible dependency versions

Resolves security audit failures while maintaining enhanced infrastructure features.
