+++ tmp/sc-network-types-patch/Cargo.toml
@@ -4,9 +4,20 @@ version = "0.15.5"
 edition = "2021"
 
 [dependencies]
-array-bytes = "6.2.3"
-codec = { package = "parity-scale-codec", version = "3.6.12", default-features = false, features = ["derive"] }
-libp2p = "0.54.1"
-multiaddr = "0.18.2"
-multihash = "0.19.3"
-scale-info = { version = "2.11.3", default-features = false, features = ["derive"] }
+bs58 = { version = "0.5.0", default-features = false, features = ["alloc"] }
+bytes = { version = "1.4.0", default-features = false }
+ed25519-dalek = { version = "2.0.0", default-features = false, features = ["rand_core"] }
+libp2p-identity = { version = "0.2.9", default-features = false }
+libp2p-kad = { version = "0.46.0", default-features = false }
+litep2p = { version = "0.9.0", default-features = false }
+log = { version = "0.4.17", default-features = false }
+multiaddr = { version = "0.18.0", default-features = false }
+multihash = { version = "0.19.0", default-features = false }
+parity-scale-codec = { version = "3.6.0", default-features = false, features = ["derive"] }
+rand = { version = "0.8.5", default-features = false }
+thiserror = { version = "1.0.0", default-features = false }
+zeroize = { version = "1.8.0", default-features = false }
+
+[features]
+default = []
+std = []
