+++ tmp/sc-network-types-patch/src/lib.rs
@@ -1,71 +1,72 @@
-// Copyright (C) Parity Technologies (UK) Ltd.
-// SPDX-License-Identifier: GPL-3.0-or-later WITH Classpath-exception-2.0
+// Re-export essential types from dependencies
+pub use libp2p_identity::PeerId;
+pub use multiaddr::Multiaddr;
 
-//! Substrate network types.
-
-pub mod kad;
-pub use multiaddr;
-
-pub use multiaddr::{Multiaddr, PeerId};
-
-// Helper function for PeerId conversions to avoid orphan rule violations
-pub fn peer_id_from_bytes(bytes: &[u8]) -> Option<PeerId> {
-    PeerId::from_bytes(bytes).ok()
-}
-
-pub fn peer_id_to_bytes(peer_id: &PeerId) -> Vec<u8> {
-    peer_id.to_bytes()
-}
-
-/// Build multiaddr from components
-pub fn build_multiaddr(peer_id: PeerId, port: u16) -> Multiaddr {
-    format!("/ip4/127.0.0.1/tcp/{}/p2p/{}", port, peer_id)
-        .parse()
-        .expect("Valid multiaddr")
-}
-
-/// Ed25519 public key utilities
-pub mod ed25519 {
-    use codec::{Decode, Encode};
-    use scale_info::TypeInfo;
+// Kad module with required types for sc-network compatibility
+pub mod kad {
+    use super::*;
     
-    /// Ed25519 public key
-    #[derive(Clone, Debug, PartialEq, Eq, Encode, Decode, TypeInfo)]
-    pub struct Public(pub [u8; 32]);
+    /// Kademlia key type - simplified implementation
+    #[derive(Debug, Clone, PartialEq, Eq, Hash)]
+    pub struct Key(pub Vec<u8>);
     
-    impl From<[u8; 32]> for Public {
-        fn from(bytes: [u8; 32]) -> Self {
-            Public(bytes)
+    impl Key {
+        pub fn new(data: Vec<u8>) -> Self {
+            Self(data)
         }
-    }
-    
-    impl AsRef<[u8]> for Public {
-        fn as_ref(&self) -> &[u8] {
+        
+        pub fn to_vec(&self) -> Vec<u8> {
+            self.0.clone()
+        }
+        
+        pub fn as_bytes(&self) -> &[u8] {
             &self.0
         }
     }
-}
-
-/// DHT record
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct DhtRecord {
-    /// Key of the record.
-    pub key: Vec<u8>,
-    /// Value of the record.
-    pub value: Vec<u8>,
-    /// Publisher of the record.
-    pub publisher: Option<Vec<u8>>,
-    /// Time-to-live of the record.
-    pub expires: Option<std::time::Instant>,
-}
-
-/// The parameters to pass to `Discovery::new`.
-#[derive(Clone, Debug)]
-pub struct DiscoveryConfig {
-    /// Identity of the local node.
-    pub local_peer_id: Vec<u8>,
-    /// TTL of the records put in the DHT.
-    pub ttl: std::time::Duration,
-    /// Interval to publish our identification.
-    pub publication_interval: std::time::Duration,
+    
+    impl From<Vec<u8>> for Key {
+        fn from(data: Vec<u8>) -> Self {
+            Self(data)
+        }
+    }
+    
+    impl From<Key> for Vec<u8> {
+        fn from(key: Key) -> Self {
+            key.0
+        }
+    }
+    
+    /// Kademlia record type - simplified implementation
+    #[derive(Debug, Clone, PartialEq, Eq)]
+    pub struct Record {
+        pub key: Key,
+        pub value: Vec<u8>,
+        pub publisher: Option<Vec<u8>>, // Use Vec<u8> instead of PeerId to avoid encode/decode issues
+    }
+    
+    impl Record {
+        pub fn new(key: Key, value: Vec<u8>) -> Self {
+            Self {
+                key,
+                value,
+                publisher: None,
+            }
+        }
+    }
+    
+    /// Peer record type - simplified implementation
+    #[derive(Debug, Clone, PartialEq, Eq)]
+    pub struct PeerRecord {
+        pub peer_id: Vec<u8>, // Use Vec<u8> instead of PeerId to avoid encode/decode issues
+        pub addresses: Vec<Vec<u8>>, // Use Vec<Vec<u8>> instead of Vec<Multiaddr> to avoid encode/decode issues
+    }
+    
+    impl PeerRecord {
+        pub fn new(peer_id: Vec<u8>, addresses: Vec<Vec<u8>>) -> Self {
+            Self {
+                peer_id,
+                addresses,
+            }
+        }
+    }
 }
